% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdminPenalty.R
\name{AdminPenalty_selectByCompany}
\alias{AdminPenalty_selectByCompany}
\title{行政处罚按公司和日期查询}
\usage{
AdminPenalty_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
行政处罚按公司和日期查询
}
\examples{
AdminPenalty_selectByCompany()
}
