% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdminPenalty.R
\name{AdminPenalty_selectByDate}
\alias{AdminPenalty_selectByDate}
\title{行政处罚按日期查询}
\usage{
AdminPenalty_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
行政处罚按日期查询
}
\examples{
AdminPenalty_selectByDate()
}
