% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankInfo.R
\name{BankInfo_selectByCompany}
\alias{BankInfo_selectByCompany}
\title{开票信息按公司和日期查询}
\usage{
BankInfo_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
开票信息按公司和日期查询
}
\examples{
BankInfo_selectByCompany()
}
