% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankInfo.R
\name{BankInfo_selectByDate}
\alias{BankInfo_selectByDate}
\title{开票信息按日期查询}
\usage{
BankInfo_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
开票信息按日期查询
}
\examples{
BankInfo_selectByDate()
}
