% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bankruptcy.R
\name{Bankruptcy_selectByCompany}
\alias{Bankruptcy_selectByCompany}
\title{破产重整按公司和日期查询}
\usage{
Bankruptcy_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
破产重整按公司和日期查询
}
\examples{
Bankruptcy_selectByCompany()
}
