% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bankruptcy.R
\name{Bankruptcy_selectByDate}
\alias{Bankruptcy_selectByDate}
\title{破产重整按日期查询}
\usage{
Bankruptcy_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
破产重整按日期查询
}
\examples{
Bankruptcy_selectByDate()
}
