% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChattelMortgage.R
\name{ChattelMortgage_selectByCompany}
\alias{ChattelMortgage_selectByCompany}
\title{动产抵押按公司和日期查询}
\usage{
ChattelMortgage_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
动产抵押按公司和日期查询
}
\examples{
ChattelMortgage_selectByCompany()
}
