% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChattelMortgage.R
\name{ChattelMortgage_selectByDate}
\alias{ChattelMortgage_selectByDate}
\title{动产抵押按日期查询}
\usage{
ChattelMortgage_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
动产抵押按日期查询
}
\examples{
ChattelMortgage_selectByDate()
}
