% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{Core_selectByCompany}
\alias{Core_selectByCompany}
\title{基本信息按公司和日期查询}
\usage{
Core_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
基本信息按公司和日期查询
}
\examples{
Core_selectByCompany()
}
