% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{Core_selectByDate}
\alias{Core_selectByDate}
\title{基本信息按日期查询}
\usage{
Core_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
基本信息按日期查询
}
\examples{
Core_selectByDate()
}
