% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stat.R
\name{count_stat_selectByCompany}
\alias{count_stat_selectByCompany}
\title{企业征信统计表按公司和日期查询}
\usage{
count_stat_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
企业征信统计表按公司和日期查询
}
\examples{
count_stat_selectByCompany()
}
