% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stat.R
\name{count_stat_selectByDate}
\alias{count_stat_selectByDate}
\title{企业征信统计表按日期查询}
\usage{
count_stat_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
企业征信统计表按日期查询
}
\examples{
count_stat_selectByDate()
}
