% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreditCalculation.R
\name{CreditCalculation_selectByCompany}
\alias{CreditCalculation_selectByCompany}
\title{信用测算按公司和日期查询}
\usage{
CreditCalculation_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
信用测算按公司和日期查询
}
\examples{
CreditCalculation_selectByCompany()
}
