% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreditCalculation.R
\name{CreditCalculation_selectByDate}
\alias{CreditCalculation_selectByDate}
\title{信用测算按日期查询}
\usage{
CreditCalculation_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
信用测算按日期查询
}
\examples{
CreditCalculation_selectByDate()
}
