% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmployeeList.R
\name{EmployeeList_selectByCompany}
\alias{EmployeeList_selectByCompany}
\title{主要人员按公司和日期查询}
\usage{
EmployeeList_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
主要人员按公司和日期查询
}
\examples{
EmployeeList_selectByCompany()
}
