% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmployeeList.R
\name{EmployeeList_selectByDate}
\alias{EmployeeList_selectByDate}
\title{主要人员按日期查询}
\usage{
EmployeeList_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
主要人员按日期查询
}
\examples{
EmployeeList_selectByDate()
}
