% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvPunishment.R
\name{EnvPunishment_selectByCompany}
\alias{EnvPunishment_selectByCompany}
\title{环保处罚按公司和日期查询}
\usage{
EnvPunishment_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
环保处罚按公司和日期查询
}
\examples{
EnvPunishment_selectByCompany()
}
