% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvPunishment.R
\name{EnvPunishment_selectByDate}
\alias{EnvPunishment_selectByDate}
\title{环保处罚按日期查询}
\usage{
EnvPunishment_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
环保处罚按日期查询
}
\examples{
EnvPunishment_selectByDate()
}
