% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquityFreeze.R
\name{EquityFreeze_selectByCompany}
\alias{EquityFreeze_selectByCompany}
\title{股权冻结按公司和日期查询}
\usage{
EquityFreeze_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
股权冻结按公司和日期查询
}
\examples{
EquityFreeze_selectByCompany()
}
