% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquityFreeze.R
\name{EquityFreeze_selectByDate}
\alias{EquityFreeze_selectByDate}
\title{股权冻结按日期查询}
\usage{
EquityFreeze_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
股权冻结按日期查询
}
\examples{
EquityFreeze_selectByDate()
}
