% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquityPledge.R
\name{EquityPledge_selectByCompany}
\alias{EquityPledge_selectByCompany}
\title{股权出质按公司和日期查询}
\usage{
EquityPledge_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
股权出质按公司和日期查询
}
\examples{
EquityPledge_selectByCompany()
}
