% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquityPledge.R
\name{EquityPledge_selectByDate}
\alias{EquityPledge_selectByDate}
\title{股权出质按日期查询}
\usage{
EquityPledge_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
股权出质按日期查询
}
\examples{
EquityPledge_selectByDate()
}
