% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exception.R
\name{Exception_selectByCompany}
\alias{Exception_selectByCompany}
\title{经营异常按公司和日期查询}
\usage{
Exception_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
经营异常按公司和日期查询
}
\examples{
Exception_selectByCompany()
}
