% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exception.R
\name{Exception_selectByDate}
\alias{Exception_selectByDate}
\title{经营异常按日期查询}
\usage{
Exception_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
经营异常按日期查询
}
\examples{
Exception_selectByDate()
}
