% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvestmentList.R
\name{InvestmentList_selectByCompany}
\alias{InvestmentList_selectByCompany}
\title{对外投资按公司和日期查询}
\usage{
InvestmentList_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
对外投资按公司和日期查询
}
\examples{
InvestmentList_selectByCompany()
}
