% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JudicialSale.R
\name{JudicialSale_selectByCompany}
\alias{JudicialSale_selectByCompany}
\title{司法拍卖按公司和日期查询}
\usage{
JudicialSale_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
司法拍卖按公司和日期查询
}
\examples{
JudicialSale_selectByCompany()
}
