% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JudicialSale.R
\name{JudicialSale_selectByDate}
\alias{JudicialSale_selectByDate}
\title{司法拍卖按日期查询}
\usage{
JudicialSale_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
司法拍卖按日期查询
}
\examples{
JudicialSale_selectByDate()
}
