% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartnerList.R
\name{PartnerList_selectByCompany}
\alias{PartnerList_selectByCompany}
\title{股东信息按公司和日期查询}
\usage{
PartnerList_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
股东信息按公司和日期查询
}
\examples{
PartnerList_selectByCompany()
}
