% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartnerList.R
\name{PartnerList_selectByDate}
\alias{PartnerList_selectByDate}
\title{股东信息按日期查询}
\usage{
PartnerList_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
股东信息按日期查询
}
\examples{
PartnerList_selectByDate()
}
