% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriousIllegal.R
\name{SeriousIllegal_selectByCompany}
\alias{SeriousIllegal_selectByCompany}
\title{严重违法按公司和日期查询}
\usage{
SeriousIllegal_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
严重违法按公司和日期查询
}
\examples{
SeriousIllegal_selectByCompany()
}
