% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriousIllegal.R
\name{SeriousIllegal_selectByDate}
\alias{SeriousIllegal_selectByDate}
\title{严重违法按日期查询}
\usage{
SeriousIllegal_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
严重违法按日期查询
}
\examples{
SeriousIllegal_selectByDate()
}
