% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiXin.R
\name{ShiXin_selectByCompany}
\alias{ShiXin_selectByCompany}
\title{失信被执行人按公司和日期查询}
\usage{
ShiXin_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
失信被执行人按公司和日期查询
}
\examples{
ShiXin_selectByCompany()
}
