% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiXin.R
\name{ShiXin_selectByDate}
\alias{ShiXin_selectByDate}
\title{失信被执行人按日期查询}
\usage{
ShiXin_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
失信被执行人按日期查询
}
\examples{
ShiXin_selectByDate()
}
