% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumptuary.R
\name{Sumptuary_selectByCompany}
\alias{Sumptuary_selectByCompany}
\title{限制高消费按公司和日期查询}
\usage{
Sumptuary_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
限制高消费按公司和日期查询
}
\examples{
Sumptuary_selectByCompany()
}
