% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumptuary.R
\name{Sumptuary_selectByDate}
\alias{Sumptuary_selectByDate}
\title{限制高消费按日期查询}
\usage{
Sumptuary_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
限制高消费按日期查询
}
\examples{
Sumptuary_selectByDate()
}
