% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{Task_selectByCompany}
\alias{Task_selectByCompany}
\title{API查询任务按公司和日期查询}
\usage{
Task_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
API查询任务按公司和日期查询
}
\examples{
Task_selectByCompany()
}
