% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxAbnormal.R
\name{TaxAbnormal_selectByCompany}
\alias{TaxAbnormal_selectByCompany}
\title{税务非正常户按公司和日期查询}
\usage{
TaxAbnormal_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
税务非正常户按公司和日期查询
}
\examples{
TaxAbnormal_selectByCompany()
}
