% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxAbnormal.R
\name{TaxAbnormal_selectByDate}
\alias{TaxAbnormal_selectByDate}
\title{税务非正常户按日期查询}
\usage{
TaxAbnormal_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
税务非正常户按日期查询
}
\examples{
TaxAbnormal_selectByDate()
}
