% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxCreditList.R
\name{TaxCreditList_selectByCompany}
\alias{TaxCreditList_selectByCompany}
\title{纳税信用等级按公司和日期查询}
\usage{
TaxCreditList_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
纳税信用等级按公司和日期查询
}
\examples{
TaxCreditList_selectByCompany()
}
