% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxCreditList.R
\name{TaxCreditList_selectByDate}
\alias{TaxCreditList_selectByDate}
\title{纳税信用等级按日期查询}
\usage{
TaxCreditList_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
纳税信用等级按日期查询
}
\examples{
TaxCreditList_selectByDate()
}
