% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxIllegal.R
\name{TaxIllegal_selectByCompany}
\alias{TaxIllegal_selectByCompany}
\title{税收违法按公司和日期查询}
\usage{
TaxIllegal_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
税收违法按公司和日期查询
}
\examples{
TaxIllegal_selectByCompany()
}
