% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxIllegal.R
\name{TaxIllegal_selectByDate}
\alias{TaxIllegal_selectByDate}
\title{税收违法按日期查询}
\usage{
TaxIllegal_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
税收违法按日期查询
}
\examples{
TaxIllegal_selectByDate()
}
