% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxOweNotice.R
\name{TaxOweNotice_selectByCompany}
\alias{TaxOweNotice_selectByCompany}
\title{欠税公告按公司和日期查询}
\usage{
TaxOweNotice_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
欠税公告按公司和日期查询
}
\examples{
TaxOweNotice_selectByCompany()
}
