% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxOweNotice.R
\name{TaxOweNotice_selectByDate}
\alias{TaxOweNotice_selectByDate}
\title{欠税公告按日期查询}
\usage{
TaxOweNotice_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
欠税公告按日期查询
}
\examples{
TaxOweNotice_selectByDate()
}
