% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZhiXing.R
\name{ZhiXing_selectByCompany}
\alias{ZhiXing_selectByCompany}
\title{被执行人按公司和日期查询}
\usage{
ZhiXing_selectByCompany(dms_token, FCompanyName, FDate)
}
\arguments{
\item{FCompanyName}{#' @param FDate}
}
\value{
无返回值
}
\description{
被执行人按公司和日期查询
}
\examples{
ZhiXing_selectByCompany()
}
