% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZhiXing.R
\name{ZhiXing_selectByDate}
\alias{ZhiXing_selectByDate}
\title{被执行人按日期查询}
\usage{
ZhiXing_selectByDate(dms_token, FDate)
}
\arguments{
\item{dms_token}{#' @param FDate}
}
\value{
无返回值
}
\description{
被执行人按日期查询
}
\examples{
ZhiXing_selectByDate()
}
